+++ src/main/java/toni/distractionfreerecipes/foundation/JEIAccessor.java
@@ -1,0 +1,9 @@
+package toni.distractionfreerecipes.foundation;
+
+import mezz.jei.common.util.ImmutableRect2i;
+
+public interface JEIAccessor {
+
+
+    public boolean distractionFreeRecipes$isSearchEmpty();
+}