+++ src/main/java/toni/distractionfreerecipes/mixins/EMIEditBoxMixin.java
@@ -6,6 +6,7 @@ import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.FastColor;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -17,7 +18,7 @@ public class EMIEditBoxMixin {
     @Redirect(method = "renderWidget", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/GuiGraphics;blitSprite(Lnet/minecraft/resources/ResourceLocation;IIII)V"))
     private void beforeRender(GuiGraphics instance, ResourceLocation resourceLocation, int x, int y, int w, int h) {
         if ((Object) this instanceof EmiSearchWidget) {
-            if (AllConfigs.client().enabled.get() && AllConfigs.client().lowerOpacity.get())
+            if (AllConfigs.client().lowerOpacity.get())
             {
                 var color = FastColor.ARGB32.color(70, 0, 0, 0);
                 instance.fill(x, y, x + w, y + h, color);