+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-## Changelog
+## 1.2.0
+- Added hide/show recipes button support
+- Added help text support to EMI
+
+## 1.1.0
+- Added JEI and REI support
+- Added option for help text rendering over the empty search list
+
+## 1.0.0
 - Initial release